requirements.txt
streamlit>=1.32,<2.0
pillow>=10.0,<11.0
google-generativeai>=0.5,<0.6
